requirements.txt
aiosqlite>=0.19.0
httpx>=0.25.0
irc>=20.3.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
asyncio>=3.4.3
typing>=3.7.4.3
uuid>=1.30
# For type hints
types-aiofiles>=23.2.0
types-python-dateutil>=2.8.19.14
mypy>=1.7.0